Hookup favorite button and delete user-library component
