--- app/templates/components/media/user-library.hbs
@@ -1,16 +1,0 @@
-{{! review }}
-{{#if (not-eq entry.progress 0)}}
-  <button class="button button--primary" onclick={{action "openReview"}}>
-    {{#if (or (not entry.review) entry.review.isNew)}}
-      {{t "media.show.sidebar.review.write"}}
-    {{else}}
-      {{t "media.show.sidebar.review.edit"}}
-    {{/if}}
-  </button>
-  {{#if reviewOpen}}
-    {{to-elsewhere named="modal" send=(component "modals/media-review"
-      onClose=(toggle-action "reviewOpen" this)
-      review=entry.review.content
-    )}}
-  {{/if}}
-{{/if}}
